create hook query params for package/name route
